+++ adodb.inc.php
@@ -198,7 +198,7 @@ if (!defined('_ADODB_LAYER')) {
 		/**
 		 * ADODB version as a string.
 		 */
-		$ADODB_vers = 'v5.22.0-dev  Unreleased';
+		$ADODB_vers = 'v5.23.0-dev  Unreleased';
 
 		/**
 		 * Determines whether recordset->RecordCount() is used.
@@ -512,35 +512,44 @@ if (!defined('_ADODB_LAYER')) {
 	/*****************************************
 	* memcached server options
 	******************************************/
-	/*
-	 * Should we use memCache instead of caching in files
+
+	/**
+	 * Use memCache library instead of caching in files.
+	 * @var bool $memCache
 	 */
 	public $memCache = false;
-	/*
-	 * A string, array of hosts or array of memcache connection
-	 * options (see adodb.org)
+
+	/**
+	 * The memcache server(s) to connect to. Can be defined as:
+	 * - a single host name/ip address
+	 * - a list of hosts/ip addresses
+	 * - an array of server connection data (weighted server groups).
+	 * @link https://adodb.org/dokuwiki/doku.php?id=v5:userguide:memcached
+	 * @var string|array $memCacheHost
 	 */
 	public $memCacheHost;
 
-	/*
-	 * Default port, may be ignored if connection object array
-	 * is set
+	/**
+	 * Default port number.
+	 * The default port can be overridden if memcache server connection data
+	 * is provided as an array {@see $memCacheHost}.
+	 * @var int $memCachePort
 	 */
 	public $memCachePort = 11211;
 
-	/*
-	 * Use 'true' to store the item compressed
-	 * uses zlib, Direct option for memcache, else
-	 * For memcached, use the memcacheOptions feature
+	/**
+	 * Enable compression of stored items.
+	 * @var bool $memCacheCompress
 	 */
 	public $memCacheCompress = false;
 
-	/*
-	 * If using mecached, an array of options
+	/**
+	 * An array of memcached options.
+	 * Only used with memcached; memcache ignores this setting.
 	 * @link https://www.php.net/manual/en/memcached.constants.php
+	 * @var array $memCacheOptions
 	 */
 	public $memCacheOptions = array();
-
 
 	var $sysDate = false; /// name of function that returns the current date
 	var $sysTimeStamp = false; /// name of function that returns the current timestamp
@@ -1022,6 +1031,8 @@ if (!defined('_ADODB_LAYER')) {
 	 *
 	 * @param string $fmt Format string
 	 * @param string $col Date column; use system date if not specified.
+	 *
+	 * @return string
 	 */
 	function SQLDate($fmt, $col = '') {
 		if (!$col) {
@@ -3911,10 +3922,21 @@ class ADORecordSet implements IteratorAggregate {
 
 
 	/**
-	 * @var ADOFieldObject[] Field metadata cache
-	 * @see fieldTypesArray()
+	* @var ADOFieldObject[] Field metadata cache
+	* @see fieldTypesArray()
 	 */
 	protected $fieldObjectsCache;
+	
+	/**
+	* @var bool True if we have retrieved the fields metadata
+	*/
+	protected $fieldObjectsRetrieved = false;
+
+	/*
+	* Cross-reference the objects by name for easy access
+	*/
+	protected $fieldObjectsIndex = array();
+
 
 	/**
 	 * Constructor
@@ -5376,12 +5398,8 @@ class ADORecordSet implements IteratorAggregate {
 				break;
 
 			case 'mysql':
-				// mysql driver deprecated since 5.5, removed in 7.0
-				// automatically switch to mysqli
-				if(version_compare(PHP_VERSION, '7.0.0', '>=')) {
-					$db = 'mysqli';
-				}
-				$class = $db;
+				// mysql extension removed in PHP 7.0 - automatically switch to mysqli
+				$class = $db = 'mysqli';
 				break;
 
 			default: