+++ drivers/adodb-mssqlnative.inc.php
@@ -1053,16 +1053,7 @@ class ADORecordset_mssqlnative extends ADORecordSet {
 	var $fieldOffset = 0;
 	// _mths works only in non-localised system
 
-	/**
-	 * @var bool True if we have retrieved the fields metadata
-	 */
-	private $fieldObjectsRetrieved = false;
-
-	/*
-	* Cross-reference the objects by name for easy access
-	*/
-	private $fieldObjectsIndex = array();
-
+	
 	/*
 	 * Cross references the dateTime objects for faster decoding
 	 */